requirements.txt
pytest==7.1.3
pytest-clarity==0.3.0a0
pytest-sugar==0.9.4
fbprophet==0.7.1
yfinance==0.1.74
streamlit==0.79.0
python-dateutil==2.8.1
pandas==1.5.1
matplotlib==3.4.3
